fix(recette): typage UNION personal_records (null::int / null::numeric)
Section 9 de recette-dataset.sql : le UNION ALL insérait des null nus dans
reps/weight_kg → Postgres les typait en text → ERROR 42804 (column "reps" is
of type integer but expression is of type text). Casts explicites dans les 3
branches. Le bloc DO $$ étant transactionnel, l'échec n'avait rien appliqué.

+++ supabase/scripts/recette-dataset.sql
@@ -413,15 +413,15 @@ begin
     )
   )
   select gen_random_uuid(), v_user, exercise_id, 'max_weight',
-         round(sess_max, 1), null, sess_max, workout_id, finished_at
+         round(sess_max, 1), null::int, sess_max, workout_id, finished_at
   from ranked where run_max is null or sess_max > run_max
   union all
   select gen_random_uuid(), v_user, exercise_id, 'estimated_1rm',
-         round(sess_1rm, 1), null, null, workout_id, finished_at
+         round(sess_1rm, 1), null::int, null::numeric, workout_id, finished_at
   from ranked where run_1rm is null or sess_1rm > run_1rm
   union all
   select gen_random_uuid(), v_user, exercise_id, 'best_volume',
-         round(sess_vol, 1), null, null, workout_id, finished_at
+         round(sess_vol, 1), null::int, null::numeric, workout_id, finished_at
   from ranked where run_vol is null or sess_vol > run_vol;
 
   ---------------------------------------------------------------------------
